src/models/baseline/pyproject.toml
[project]
name = "baseline-service"
version = "0.1.0"
description = "Baseline perceptual hash service using perception library"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "Pillow>=10.0.0",
    "tqdm>=4.65.0",
    "perception>=0.7.7",
    "faiss-cpu>=1.7.4",
    "tools",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
tools = { path = "common/tools-0.1.8-py3-none-any.whl" }
